// types/global.ts
export interface Paginate<T> {
  data: T[];
  links: {
    first: string;
    last: string;
    next: string | null;
    prev: string | null;
  };
  meta: {
    current_page: number;
    from: number;
    last_page: number;
    path: string;
    links: {
      active: boolean;
      label: string;
      url: string;
    }[];
    per_page: string;
    to: number;
    total: number;
  };
}
// ...
export type LikeTypes = "product" | "shop" | "master";

export interface LikeOptions {
  type: LikeTypes;
  type_id?: number;
  region_id?: number;
  country_id?: number;
  city_id?: number;
}